Implemented template list
Implemented type-safe, template doubly-linked list in C.
Added the most necessary operations for modifying the list,
iterating over it and finding elements in it.
The list also features iterator concept for denoting elements
and traversing the list.

All the implementation details are hidden behind the convenient macros.

The demo program has been created to demonstrate the ease of use
and beauty of the template list.